feat: move video call page to general index
chore: deprecate old index
